refactor: Remove Sub form components
