+++ packages/ui/src/forum/components/category/index.ts
@@ -1,2 +1,1 @@
 export * from './ForumCategoryList'
-export * from './SubForumCategoryList'
